refactor: 버그수정 및 로직변경
