+++ next.config.ts
@@ -4,7 +4,7 @@ const nextConfig: NextConfig = {
   reactStrictMode: false,
 
   images: {
-    domains: ["imgnews.pstatic.net", "54.180.104.178"],
+    domains: ["imgnews.pstatic.net", "43.201.99.157"],
     remotePatterns: [
       {
         protocol: "http",
